Rename GetAvailableRooms with GetRooms
