+++ Vonage.Test.Unit/Meetings/GetRooms/Data/ShouldDeserialize200-response.json
